@GitHK review: add test
